Moved all providers to the same package
